Gallery execution + plotly parity + universal loader (round 6)
Gallery: sphinx-gallery previously executed only plot_*-named examples,
leaving chemtrails/animate*/precog/explore/save_*/analyze pages with
code but no rendered output. All examples now execute
(filename_pattern), matplotlib animations render as embedded mp4 video
(matplotlib_animations + sphinxcontrib-video; animation examples expose
`ani = ani_geo.line_ani` for the scraper), save_* examples write to
temp files, and every example page gets a branch-aware "Open in Colab"
badge + .ipynb link (post_build.py) so gallery examples open as
runnable notebooks.

Dual-backend audit (scripts/audit_gallery_backends.py): every example
runs under BOTH matplotlib and plotly in subprocesses; 78/78 pass
(save_movie under plotly needs a long timeout -- kaleido per-frame mp4
export). The audit caught a real, years-latent bug: per-dataset fmt
lists (['-','--']) routed each dataset through interp_array_list
(plural), silently replacing 2D arrays with lists of per-row
interpolations; latent because is_line() always returned False before
its round-2 fix. Fixed with interp_array + regression test.

Plotly parity for animation extras: chemtrails/precog/bullettime draw
low-opacity trail traces on window animations, tail_duration sets the
window length, and zoom moves the camera (r = 1.95*(9-zoom)/8,
mirroring the matplotlib zoom semantics); previously none of these were
forwarded to the plotly renderer. explore maps to plotly's native
hover.

Universal loader: hyp.load (and DataGeometry.plot/transform) resolve
strings by trying, in order: built-in dataset name -> local file
(npy/npz/csv/tsv/txt/json/parquet/mat/pickle) -> Hugging Face dataset
(split=/streaming=; streaming feeds straight into hyp.plot) -> Google
Drive URL or bare id -> Dropbox URL or shared-link path -> any URL with
or without https://. Lists of strings return lists of datasets. Raw
text (whitespace) still flows to the text-embedding pipeline. Also
fixes df2mat for pandas>=2 (get_dummies bool dtype made mixed
DataFrames produce object arrays that crashed np.isnan).

19 new tests (12 loader incl. real HF/Drive/Dropbox/URL fetches, 6
plotly trails, 1 interp regression); 232 passing.

+++ hypertools/tools/df2mat.py
@@ -33,7 +33,11 @@ def df2mat(data, return_labels=False):
     df_num = data.select_dtypes(exclude=['object'])
 
     for colname in df_str.columns:
-        df_num = df_num.join(pd.get_dummies(data[colname], prefix=colname))
+        # dtype=float: pandas >= 2.0 defaults get_dummies to bool, and
+        # joining bool dummies with float columns makes .values an
+        # object-dtype array that np.isnan / reducers cannot handle
+        df_num = df_num.join(pd.get_dummies(data[colname], prefix=colname,
+                                            dtype=float))
 
     plot_data = df_num.values
